# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event Manager",
    "description": "Event Manager is an application where users who are managed by admin can create events and choose a center where they want the event to take place",
    "version": "1.0",
    "contact": "email: kenolusola@gmail.com"
  },
  "servers": [
    {
      "url": "https://emanager1980.herokuapp.com",
      "variables": {}
    }
  ],
  "paths": {
...
    "/api/v1/centers": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get All Centers",
        "description": "API endpoint to get all centers",
        "operationId": "GetAllCenters",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully gets all centers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Success",
                    "message": "List of all Centers"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Centers not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "No centers found"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Server Error"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create Center",
        "description": "API route that allows Superadmin and Admin create a center",
        "operationId": "CreateCenter",
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "application/x-www-form-urlencoded"
            },
            "example": "application/x-www-form-urlencoded"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "name",
                  "location",
                  "description",
                  "capacity",
                  "facilities",
                  "image",
                  "price"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Emerald Enclave"
                  },
                  "location": {
                    "type": "string",
                    "example": "6, Aminu street Garki Abuja"
                  },
                  "description": {
                    "type": "string",
                    "example": "For your eyes only"
                  },
                  "capacity": {
                    "type": "string",
                    "example": "200"
                  },
                  "facilities": {
                    "type": "string",
                    "example": "swimming pool"
                  },
                  "image": {
                    "type": "string",
                    "example": "\"\""
                  },
                  "price": {
                    "type": "string",
                    "example": "10000"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": "Successfully created a center",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Success",
                    "message": "Center created successfully"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Server Error"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
    "/api/v1/centers/{centerId}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get One Center",
        "description": "API endpoint to get one center",
        "operationId": "GetOneCenter",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully gets one center",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Success",
                    "message": "List of one center"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Center not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "No center found"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Server Error"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Modify Center",
        "description": "API endpoint to modify center by SuperAdmin or an Admin",
        "operationId": "ModifyCenter",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwNzcxNywiZXhwIjoxNTI2NTgwNTE3fQ.D2OiADSZWRBk4lZDp0Ys3nr5i354x8wfMqugKkXTJt8"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwNzcxNywiZXhwIjoxNTI2NTgwNTE3fQ.D2OiADSZWRBk4lZDp0Ys3nr5i354x8wfMqugKkXTJt8"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "application/x-www-form-urlencoded"
            },
            "example": "application/x-www-form-urlencoded"
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "name",
                  "location",
                  "description",
                  "capacity",
                  "facilities",
                  "image",
                  "price"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "The Gooche"
                  },
                  "location": {
                    "type": "string",
                    "example": "6, Aminu street Garki Abuja"
                  },
                  "description": {
                    "type": "string",
                    "example": "For your eyes only"
                  },
                  "capacity": {
                    "type": "string",
                    "example": "1000"
                  },
                  "facilities": {
                    "type": "string",
                    "example": "swimming pool"
                  },
                  "image": {
                    "type": "string",
                    "example": "\"\""
                  },
                  "price": {
                    "type": "string",
                    "example": "20000"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Successfully modified one center",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Success",
                    "message": "Center updated"
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authorised to modify center",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "User cannot modify this center"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Center not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Center not found"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Server Error"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete Center",
        "description": "API endpoint to delete centers",
        "operationId": "DeleteCenter",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwNzcxNywiZXhwIjoxNTI2NTgwNTE3fQ.D2OiADSZWRBk4lZDp0Ys3nr5i354x8wfMqugKkXTJt8"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwNzcxNywiZXhwIjoxNTI2NTgwNTE3fQ.D2OiADSZWRBk4lZDp0Ys3nr5i354x8wfMqugKkXTJt8"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Success",
                    "message": "Center has been successfully deleted"
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authorised to delete center",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "User cannot delete this center"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Center not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Center not found"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Server Error"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
...
  }
}